Parse version banner from clients.
RFCs is a bit permissive here, it allows custom lines to be sent before the SSH-
line, it also requires that SSH- starts in a new line.

I've tried to be very careful and consider every case.

I'm still unsure on how to do the stream processing in a side effect free style
that doesn't suck, I'll stick with this so I can run tests until I know what to
do, in other words, this code will be rewritten over and over.

Still untested code, but enough for writing regression tests.