Add the begginings of Server.
Handling SSH_KEX_INIT and SSH_KEXDH_INIT is more or less done on the server
side. We will be able to unify most of the code with a client later.

First state transactions on Server.t, the idea is that input_msg() does all
Server.t transitions, so we never pass it down.

Code assumes Rsa as host_key and so on, I think it is premature to abstract at
this point, as we add more algorithms we add more abstractions.