Save derived keys in Server.t.
When the server gets a KEXDH_INIT and assembles a KEXDH_REPLY, we have
everything we need to derive the new keys, but we can't really use them until we
send a NEWKEYS message.

The plan is to return two messages when we get a KEXDH_INIT, a KEXDH_REPLY,
_and_ a NEWKEYS.

Then, the caller of input_msg will "register a send" on each of them, something
like output_msg(), when output_msg() sees a NEWKEYS, it swaps the
Server.t.new_keys with keys.