Integrade Decode.get_pkt and Crypto.decrypt.
This changes the api of Crypto.hmac as well, fuck it, it's a pain in the ass to
make the caller include the sequence number, so pay the price of the extra
allocation and use new Nocrypto hmacv function.

There are still serious problems with this approach as we decrypt a whole buffer
even though we only look at one packet.