Preliminary channels implementation, handle open request.
First, plug back Raw_data in Ssh.channels_data, kmerz@ was right. We need to be
able to serialize an unknown channel data so that we can fail gracefully when
the protocol handles it, this is not a hard error and we just reject that
channel open request.

Channels have two endpoints, local and remote, they are keyed by an int32 id and
express window and packet limits for each endpoint.

I've implemented channels as a Map, since we have to be pure.
I'm allowing 1024 channels per connection, the allocation code is naive and just
selects the last id + 1, if the id is = 1024 we start a linear search, which
attempts to find a free key slot by traversing the map.

OpenSSH accepts our confirmation reply and we can start playing ball.
Tests must be added for all cases.