Merge Channels into Channel.db.
I never know if I should use a submodule or not. Everytime I attempt to play
with submodules, I regret and go back to having a flat Module with a "theme".

So Channel.t is a channel, and Channel.db is the database for Channel.t, same
thing in Auth.