Experiment more with the API.
It makes no sense to give the caller the resposability to accept channel
requests or not. For now, accept them all, in the future, accept a callback or a
list of acceptable requests.

So far the best solution for the replies X events is to split them, this makes
the caller considerably simple, at the cost of having a 3 sided tuple.