Upper API (Driver) is not pure anymore.
After much consideration, I'm leaning towards having the protocol implementation
in a purely functional style, but having the module that "drives" the protocol
to be non pure.

A plethora of problems arise from a pure Driver, it complicates the API, makes
it error prone for users and would perform worse since we were relying on too
much queuing of jobs/results.

The Driver is now way simpler than pure Server.Engine, although it is highly
experimental and I expect to change it a lot.

In the future, I'd like to have Driver to be an interface specification, proving
some simple ones like Unix, Lwt, Mirage and so on.