Respect peer window when sending channel data.
When we attempt to send channel data, we must respect two things:
 - Maximum packet size as advertised by peer in channel_request.
 - Available window as advertised by peer.

Channel.output_data makes all the cool stuff, and returns us a list of
Ssh.Message_channel_data. If there is not enough window space, the remaining
data is kept in Channel.tosend.

Code that processes Channel.tosend is in the oven.