Make test framework return a result.
This should require a bunch of test rewrites, but nothing breaks with this,
exceptions are converted to errors and everyone is happy.

The reason for doing this is that the test code read "too" different than the
real code, since one was relying on result binding, while the other on get_ok.

I'm neurotic and stuff should read always the same.