Awa_lwt is here \o/
This should be the main driver people use, at this time it's using Lwt_unix
directly. This will either be functorized, so Mirage can use it, or duplicated
into Awa_mirage, still unclear which is the path to go.

With Lwt we can solve most of the issue of blocking, respecting the window.
The use code is also ridiculously simple, here is the echo server:

let exec cmd sshin sshout ssherror =
  let rec loop ()  =
    sshin () >>= fun r -> match r with
    | `Eof -> Lwt.return_unit
    | `Data input -> sshout input >>= fun () -> loop ()
  in
  Lwt_io.printf "Executing %s\n%!" cmd >>= fun () ->
  loop () >>= fun () ->
  Lwt_io.printf "Execution of %s finished\n%!" cmd