client: support for password authentication fixes #31
This changes the API of Client.make (now receives a variant
`Password of string | `Pubkey of key), and Awa_mirage.Make().client_of_flow as
well.

This does not support the keyboard-interactive authentication mechanism
(RFC 4256) that is commonly used (with e.g. PAM), but only the password
authentication. May be useful for feature parity between server and client, and
for some network devices.